Add initial reservoir DA functionality for NHF

+++ src/troute-network/troute/network/reservoirs/hybrid/hybrid.pxd
@@ -9,7 +9,7 @@ cdef void run_hybrid_c(_Reach* reach, float inflow, float lateral_inflow, float 
 
 cdef extern from "hybrid_structs.h":
   ctypedef struct _MC_Hybrid:
-    int lake_number
+    long lake_number
     float dam_length, area, max_depth
     float orifice_area, orifice_coefficient, orifice_elevation
     float weir_coefficient, weir_elevation, weir_length
