+++ src/troute-network/troute/network/reservoirs/hybrid/hybrid_structs.h
@@ -9,7 +9,7 @@
 #define HYBRID_MAX_START_DATE_LENGTH 20 
 
 typedef struct {
-  int lake_number;
+  long lake_number;
   float dam_length, area, max_depth;
   float orifice_area, orifice_coefficient, orifice_elevation;
   float weir_coefficient, weir_elevation, weir_length;
