+++ src/troute-network/troute/network/reservoirs/levelpool/levelpool.pyx
@@ -8,7 +8,7 @@ Externally defined symbols
 """
 
 cdef extern from "levelpool_structs.c":
-  void init_levelpool_reach(_Reach* reach, int lake_number,
+  void init_levelpool_reach(_Reach* reach, long lake_number,
                             float dam_length, float area, float max_depth,
                             float orifice_area, float orifice_coefficient, float orifice_elevation,
                             float weir_coefficient, float weir_elevation, float weir_length,
